+++ tests/php-unit-tests/unitary-tests/setup/moduledependency/ModuleTest.php
